Separating commands and cli modules
