+++ profiles/factory.py
@@ -2,8 +2,8 @@ from flask import Flask
 from flask_migrate import Migrate
 
 from profiles.api import errors, oauth2, ping
+from profiles.cli import ClearCommand
 from profiles.clients import Clients
-from profiles.commands import ClearCommand
 from profiles.config import Config
 from profiles.exceptions import ClientError, OAuth2Error
 from profiles.models import db
